Add index.ts for the role of exporting from other files
